requirements.txt
cryptography>=43.0.3
PyQt6>=6.7.1
requests>=2.32.3
tpm2-pytss>=2.3.0
argon2-cffi>=23.1.0
pyOpenSSL>=25.0.0